backend/requirements.txt
# FastAPI and dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# OCR
pytesseract==0.3.10
Pillow==10.1.0

# Speech processing
openai-whisper==20231117
faster-whisper==0.9.0
librosa==0.10.1
parselmouth==1.1.1

# Text-to-Speech
gtts==2.4.0
coqui-tts==0.22.0

# Machine Learning
torch==2.1.1
transformers==4.35.2
accelerate==0.24.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
